Makes the binding_specs arg of new_object_graph() take instances, not classes.
Since I've decided not to auto-detect and auto-load binding specs, I don't
need to be able to create and use them from the class alone, and it seems like
it may be helpful to be able to pass in initializer params.  And then it makes
the configure() and other methods less awkward that they are non-static
methods.